Polishing and refactor

+++ solve_multiple_instances.py
@@ -3,6 +3,7 @@ import argparse
 import json
 import shutil
 import yaml
+import copy
 
 from solve_single_instance import main
 
@@ -23,17 +24,27 @@ configs_file_path = Path(args.configs).resolve()
 with open(configs_file_path, 'r') as file:
     configs = yaml.load(file, yaml.Loader)
 
-for config in configs:
+for config_name, config_changes in configs['groups'].items():
+
+    config = copy.deepcopy(configs['base'])
+    for config_changes_key, config_changes_value in config_changes.items():
+        if type(config_changes_value) is list:
+            config[config_changes_key].extend(config_changes_value)
+        else:
+            config[config_changes_key] = config_changes_value
 
     for group_directory_path in groups_input_directory_path.iterdir():
         if not group_directory_path.is_dir():
             continue
 
         group_name = group_directory_path.name
+
+        if 'groups_to_do' in config and config['groups_to_do'] != 'all' and group_name not in config['groups_to_do']:
+            continue
         
         for master_instance_file_path in group_directory_path.iterdir():
             
-            new_group_directory_path = groups_output_directory_path.joinpath(f'{group_name}_{config['name']}_{master_instance_file_path.stem}')
+            new_group_directory_path = groups_output_directory_path.joinpath(f'{group_name}_{config_name}_{master_instance_file_path.stem}')
             if new_group_directory_path.exists():
                 shutil.rmtree(new_group_directory_path)
             new_group_directory_path.mkdir()
@@ -49,4 +60,4 @@ for config in configs:
                 json.dump(master_instance, file, indent=4)
             
             print(f'Start solving instance {master_instance_file_path.name}')
-            main(new_group_directory_path, config, config['name'])
+            main(new_group_directory_path, config, config_name)
